Fix a bug with manually writing paths on browser
It is possible that a user can decide to manually enter paths to certain
routes. For intance, they can write "https:<bla blah
blah>/practitioners/me", while they have logged in as a patient, thus
the only link that should be accessible to them is "https:<bla blah
blah>/practitioners/<blah blah blah>". The back was that doing this
broke the site